requirements.txt
Flask
tensorflow
firebase-admin
Pillow
numpy
absl-py
werkzeug
requests
gunicorn